Fix use of hash value type
First of all, don't use positional parameter of role's signature where
it is used at compile-time. In particular, `is default` would then
unconditionally be set to `Mu` because this is what is the positional's
value when role body is compiled.

Second, override `of` method because the default one relies on hash own
container descriptor and knows nothing about role's value type.